fix(runtime): direct return when free a NULL pointer
This problem is triggerred in WSL + Docker, causing
`DlsymAlloc::PointerIsMine` return false.
Then XSan uses interceptor free  before InitializeMainThread (in tsan's
InitializeShadowMemory -> dlsym).
Eventually XSan gets SEGV when attempting to access XsanThread.

The temporal fix is to return directly when the ptr pending to free is NULL.

TODO: figure out WHY
- in Ubuntu + Docker env, libdl.so is preferred and
the pointer pending to free is NOT NULL in InitializeShadowMemory/dlsym.
- in WSL + Docker, libdl.a is preferred and
the pointer pending to free is NULL in InitializeShadowMemory/dlsym.