fix(test): Relax the virtual memory limit on allocator_oom_test.cpp
Adding restrictions on virtual memory in the tests is intended to
trigger out-of-memory conditions. In ASan, the protected area is small,
resulting in low virtual memory requirements. However, in XSan, the
protected area is large, and restricting virtual memory can lead to mmp
protection memory failures. Moreover, not restricting virtual memory
still allows the out-of-memory error to be correctly triggered in XSan
tests.