feat-upd(pass): Make sanitizer passes idempotent
This commit is migrated from LLVM commit 62ced81.

This commit changes the sanitizer passes to be idempotent.
When any sanitizer pass is run after it has already been run before,
double instrumentation is seen in the resulting IR. This happens because
there is no check in the pass, to verify if IR has been instrumented
before.

This commit checks if "nosanitize_*" module flag is already present
and if true, return early without running the pass again.