fix(test): The TSan itself has issues with intercepting the `clone`
The semantics of the `clone`  function vary depending on the flags
passed to it. TSan does not account for the different semantics of the
`clone` function. In future integration, the TSan `clone` interception
function will be fixed. To ensure the ASan test proceeds smoothly, the
semantics of the `clone` function in this test will be modified to align
with these mantics supported by TSan.