fix(compiler): put runtime in front of comp. cmds
Fixes ASan testcase init_fini_sections.cpp.

CTor are those functions in .pre_init, .init section or
attributed with constructor.
Link order determines the order of CTor  invokements.

We should put the sanitizer static runtime library just ahead of the
input  *.o/*.c files, so that the CTor of sanitizers can be called
before the CTors of user code (.preinit, module_ctor).

What's more, the compiler also places the sanitizer runtime library
before the input files after the Driver generates the command line,
which should be adhered by this compiler wrapper.

However, it is hard to judge whether a argument is an input file or not
precisely, so we simply add the sanitizer runtime library as front as
possible.