fix(test): update the testcase compilation option
Fixes TSan's testcase user_malloc.cpp.

This testcase registers a user `malloc` that covers XSan's one.
This user malloc would be invoked in XSan's initialization before
XSan allocates Shadow memory.

In before, we use `-xsan -fno-sanitize=thread` option to compile
the testcase, which means ASan is enabled for the user malloc.
The very early ASan check in this user malloc leads to a SEGV as the
relevant shadow has not been prepared.