feat(runtime): handle `vfork` unifiedly
Provide 2 hooks to process the `vfork` interceptors:
- `__xsan_vfork_before_and_after` for macro `COMMON_INTERCEPTOR_SPILL_AREA`
- `__xsan_vfork_parent_after` for macro `COMMON_INTERCEPTOR_HANDLE_VFORK`

For ASan, we change nothing and just migrate its unpoisoning logic to
the new hooks.

For TSan, we do not adopt its original logic that replaces the `vfork`
with `fork`. Instead, we use the unified `vfork` interceptor with some
assembly trickery, and disable TSan in vfork child, adhering the
solution discussed in the comment of TSan's vfork interceptor.