feat: intercept sub-sanitizer's init functions
ASan and TSan register their init functions in the instrumentation.

We want to use `__xsan_init` to unify the initialization of all
sanitizers when XSan is enable. However, modifying the instrumentation
code is a intrusive way.

Considering the extension, we use the following non-intrusive approach:

Use `-Wl,-wrap,<symbol>` to intercept the `__xxx_init` functions in the
link-time, and redirect them to `__xsan_init`.