feat(runtime): Unify the storage and retrieval of stacktrace in `melloc`
When storing the stack trace for `melloc` in TSan, first retrieve the
corresponding `AsanChunk` object based on the address. If the
`AsanChunk` exists, it means ASan has already stored the stack trace for
this `melloc`, and TSan will not store it again. If the `AsanChunk` does
not exist, TSan will store the stack trace for this `melloc`.

When retrieving the stack trace for `melloc` in TSan, first obtain the
corresponding `AsanChunk` object based on the address. If the
`AsanChunk` exists, it means ASan has already stored the stack trace
for this `melloc`, and TSan retrieves it through the `AsanChunk`. If the
`AsanChunk` does not exist, TSan retrieves the stack trace it stored for
this `melloc`.