refactor(runtime): modify SymbolizeData by link-time interception.
ASan modifies the global memory structure, so we need to correct the
global variable description obtained from the symbolizer.

The modification required for this feature is too small to be worth
modifying the source code specifically for this purpose,
so it is instead implemented in the form of interception when linking.