feat(pass): consider REL/ACQ barrier for ActiveAnalysis
If applied to TSan, we must consider the memory barriers.
i.e., RELEASE/ACQUIRE.
Therefore, for TSan, we extend this lattice
from
    ⊥ < ~ < ⊤
to
    ⊥ < ~ < A < ⊤
    ⊥ < ~ < R < ⊤
where A represents any acquire barrier, and R represents any release
barrier. before this program point.

If KillingI dominates DeadI (Killing --> Dead),
DeadI could not be eliminated if there is a release barrier between
them, as the DeadI is not guarded by the release barrier.
Otherwise, FN arises if other threads access the same memory location
after crossing the relevant acquire barrier.

Similarly, if KillingI post-dominate DeadI (Dead --> Killing).
DeadI could not be eliminated if there is an acquire barrier between
them, as the DeadI is not guarded by the acquire barrier.