feat(pass): support invariant check relocatiing primitively
If a MOP in loop accesses an loop-invariant address, we consider the
relevant MOP check is also loop invariant, which could be relocated into
the outer of loop.

The invariant check is, either hoisted into the preheader, or sinked
into the exit block, of the loop.

(Pending to do) If the check is in branch, we should instrument a extra
bool indicator to indicate whether the MOP is executed or not. In this
case, we can only sink the relevant check to the loop exit block.