perf(pass): transfer alloca into register
The last commit introduces some indicator to record whether the
relevant MOP is executed or not.
which should have been implemented as PHINodes in SSA form.

Nevertheless, to be simple, we just use alloca/load/store to implement
the relevant functionality.

This commit adds a `-mem2reg` pass to tranfer the instrumented alloca
into PHINode, which would promote the subsequent optimizations.