fix(pass): Derive Load Loop Invaraince
In previous `LoopInvariantChecker`, we did not specialize `LoadInst`,
leading to the wrong loop optimization.

Obviously, any operands of `LoadInst` is loop invariant cannot give us
the conclusion that `LoadInst` is loop invariant.

Besides the operands, we should also check any `StoreInst` in the same
loop to verify the value in the address must be unchanged in loop

We cannot use the unified method (i.e., Intrinsic::invariant_start) to
judge the invariance of load instructions just like what LICM does
(the relevant code is in `isLoadInvariantInLoop` @ LICM.cpp).

1. Reason 1: O0 optimization does not insert the invariant.start
intrinsic.
2. Reason 2: MSSA is incorrectly affected by calls to UBSan's functions.

Therefore, we just analyze the relevant store instruction to determine
if it is possible to write to the memory location of the load
instruction.

CallInst is ignored, because we assume that the Loop is a simple loop
without any calls that may write to memory.