fix(pass): fix type mismatch in SCEV operations
Previously, we did not consider whether the types matched when operating
 on SCEV, which could have resulted in incorrect handling of integers.

sqlite3.c triggers this issue. We multiply a i32 counter with i64
factor,  transferring a negative i32 -1 to be a positive number,
and thus leading to overflow on xsan's checking.

Now we fix it by transferring the SCEV with type PtrIntTy in the
beginning, e.g.,

previous:
```c
#define counter = ((i8) num + (i8) -1 * min(1, num))
i64 offset = (i16)2 * counter
                = 2 * (i16) num + 2 * (i16) 0xff * min(1, num)
                = 2 * (i16) num + (i16) 0x1fe * min(1, num)
```

current:
```c
#define counter = ((i16)((i8) num + (i8) -1 * min(1, num)))
i64 offset = (i16) 2 * counter
                = (i16) 2 * ((i16)((i8) num + (i8) -1 * min(1, num)))
```