feat(utils): get block address of instruction
Check whether the Instruction is on the top of the BB (ignore debug/phi
instruction).If so, just return its parent BB address. Otherwise,split
the BB on the location of this instruction, and return the new split BB
address. Besides,Delete InstUtils.cpp and InstUtils.h, Change the place
where InstUtils was originally used to ValueUtils.

fix(util): delete SplitPoint