feat(compiler): support clang installed by apt (#23)
Previously, we only used dynamic symbol overriding to implement the symbol interceptions.

However, this does not make sense for the `clang` installed via `apt`, which only relies on a shared library
`libclang-cpp.so` linked with `-Wl,-symbolc=function` parameter. Because this link parameter 
makes the dynamic function symbols used in the library itself become interior symbols, i.e., use 
relative addressing instead of addressing via *plt* table.
Note that, if we use a locally built `clang` with the default building option, it would rely on some shared 
libraries with `-symbolic` link option. I.e., the dynamic symbol overriding scheme of livepatch works.

Considering the above scenario, we determine to use a more thorough approach to implement
the livepath, i.e., rewrite the entry of specific functions to add trampoline to our interceptor functions.
E.g., for x86_64, we can replace the entry code of the given function to 
```assembly
      // movabs rax, <MyFunc>
      0x48, 0xb8, 0, 0, 0, 0, 0, 0, 0, 0,
      // jmp rax
      0xff, 0xe0
```
If our interceptor is just a wrapper on the intercepted function, we use a scoper to recover the original
code just in a given lifetime.

The new livepatch schema should also work for those unexported functions, only requiring an approach
 to obtain the addresses for them (for dynamic symbols, we can easily do that via `dlsym`)

---------

Co-authored-by: CCamsyn <2742046922@qq.com>