fix(runtime): explicitly set the symbol to default Visibility
some wrapped symbols is hidden by default
when xsan\_rtl\_static\.cpp\.o is
compiled with -fvisibility=hidden, which causes external dynamic
libraries (DSOs) to fail to find the symbol when linking, triggering an
error from the GNU ld linker. Forcing its visibility to default with
this attribute ensures that it will be visible to DSOs, and is thus
compatible with stricter linkers.