feat(livepatch): support livepatch gcc's linking
This commit removes the internal sanitizer runtime linking in gcc.
The core of this commit is a rewrite of link_command_spec

Ideally, we would need to fully patch the gcc linking process, but this
requires intercept non-dynsym function `do_spec`,
which is especially difficult after gcc is stripped

Therefore we chose to rewrite gcc's link command template --
link_command_spec, which is a constant string in rodata.
Accordingly, we can only shorten it, not lengthen it, to
avoid overwriting other strings.

Finally, considering the modification length, we decided to delete only
the sanitizer link command inside gcc in the livepatch, and let the
external wrapper complete the remaining XSan Rt link command,
unlike clang's hot patch, which does not require
 wrappers to participate).