fix: not use __cxa_guard_* (#102)
Static var in function with non-const initial value (which means lazy
initialization) is a CPP feature. It calls __cxa_guard_* in libcxx. So
when user link our lib without linking libcxx (such as `xclang -xsan
a.cpp -o a`), the linker failes. This PR avoids using them.

NOTE: TSan define a strong symbol to intercept them, so we need to turn
off TSan when building XSan to reproduce the error.