fix(build): fix DSO searching issue for gcc-patch in install scenario
If LLVMSupport is a shared library, user might get the error when using
livepatch for gcc, as follows:
```
gcc: error while loading shared libraries: libLLVMSupport.so.15:
     cannot open shared object file: No such file or directory
```

libgcc-patch.so relies on libLLVMSupport.so, and libLLVMSupport.so
 relies on libLLVMDemangle.so.
Both of libLLVMSupport.so and libLLVMDemangle.so have the RUNPATH
`$ORIGIN/../lib`.
Therefore, we can set the RUNPATH of libgcc-patch.so to `$ORIGIN/../lib`
as well, and also install the LLVMSupport and LLVMDemangle to the
`patch/../lib` directory, as libgcc-patch locates at directory `patch`.

Note:
1. Privilege: RPATH > LD_LIBRARIES_DIR > RUNPATH
2. $ORIGIN indicates that search the DSO from the directory of
 the file itself, i.e., `dirname libgcc-patch.so`