fix(ci): fix the incorrect milestone-pr-note CI (#131)
This pull request introduces updates to the release notes configuration
and workflow to enhance clarity and functionality. Key changes include
adding a structured changelog configuration and upgrading the release
notes generator action to the latest version.

### Improvements to release notes configuration:
*
[`.github/release-notes.yml`](diffhunk://#diff-4ec453fee39334f2ea10d0ca905ddadf0ebdfa6a4feadf7ea346e71a90155ef4R1-R16):
Added a structured changelog configuration with sections for major
features, minor features, and bug fixes, along with exclusion rules for
specific issue labels and contributors.

### Workflow updates:
*
[`.github/workflows/prepare-release.yml`](diffhunk://#diff-ddf2bab74923dcd43f9f8d05e50bb1adc9ee5366a394d151d1f04a54a6079486L52-R52):
Upgraded the `release-notes-generator-action` from version `2.0.1` to
`3.1.6` to take advantage of the latest features and
improvements.Upgrade the version and add missing configuration files