fix(compiler): decide livepatch or not on configure time
Previously, we used a script to dynamic determine whether to use
livepatch.

However, the reality is too complex -- we have different types of
compiler,, e.g., static-linked clang, dyn-linked clang,
local-built clang and etc...

Therefore, it is too hard to write a script that determine whether it is
appropriate to apply livepatch.

Eventually, we give the choice to user on configuration time, i.e.,
- `-DXSAN_USE_LIVEPATCH=OFF` : use invasive patch
- `-DXSAN_USE_LIVEPATCH=ON`: (default) use livepatch