fix(test): support XSAN_SYMBOLIZER_PATH in LIT (#167)
Previously, the LIT ignored the XSAN_SYMBOLIZER_PATH environment
variable, which is used by the LLVM sanitizers to locate the symbolizer
executable.

This patch adds support for this variable by passing it to the LIT
invocation.

I.e., we can use `export XSAN_SYMBOLIZER_PATH=xxx` before the LIT test
to specify the external llvm-symbolizer