fix(rt): if mem layout incompatible, ReRun w/o ASLR (#170)
This PR finally fixes #159 

This change relands an incorrect fix, see issue #169 for details.

Fix the incorrect init process which `CheckAndProtect` memory too early.

Adhering to the logic of TSan, we should disable ASLR and rerun the PUT
when the memory layout is incompatible.

Disable ASLR if the memory layout was incompatible. Alternatively, we
could just keep re-execing until we get lucky with a compatible
randomized layout, but the risk is that if it's not an ASLR-related
issue, we will be stuck in an infinite loop of re-execing (unless we
change ReExec to pass a parameter of the number of retries allowed.)