fix(compiler): only transmit linking options if -c not present (#184)
This pull request makes a small update to the `edit_params` function in
`src/compiler/clang_wrapper.c`.

The change ensures that the linker (`lld`) and additional library path
(`-L/usr/local/lib`) are only added to the compiler parameters when the
`-c` flag is not present.

- Only add `-fuse-ld=lld` and `-L/usr/local/lib` to `cc_params` if the
`-c` flag is not set, preventing unnecessary linker flags during
compilation-only steps.