doc(eval): fix some display issue in MarkDown (#190)
This pull request makes documentation improvements to the
`evaluation/README.md` file, focusing on standardizing formatting and
clarifying the directory structure. The main changes are the consistent
use of backticks for code identifiers and minor adjustments to the
example directory tree for better readability.

Formatting and clarity improvements:

* Standardized formatting by wrapping placeholders like `<package>`,
`<program1>`, and `<prog>.dbg` in backticks throughout the documentation
to clearly indicate code identifiers.
[[1]](diffhunk://#diff-f805fea08479ad131f67bfc6e9ecc475af423fb18f8c4bb760c625c88a57c495L3-R3)
[[2]](diffhunk://#diff-f805fea08479ad131f67bfc6e9ecc475af423fb18f8c4bb760c625c88a57c495L16-R38)
* Updated the example directory tree to match the described structure
and moved `activate_compile_flags.sh` to its correct position for
consistency.
[[1]](diffhunk://#diff-f805fea08479ad131f67bfc6e9ecc475af423fb18f8c4bb760c625c88a57c495L51)
[[2]](diffhunk://#diff-f805fea08479ad131f67bfc6e9ecc475af423fb18f8c4bb760c625c88a57c495R156-R160)