fix(compiler): support `xgcc -flto` now (#203)
Previously, we did not livepatch the symbols in lto mode, making `xgcc
-flto` fail to use XSan.

Now we support this scenario.

---

This pull request makes a minor update to the GCC diagnostic livepatch
code, specifically enhancing the interceptor configuration to support an
additional compiler frontend.

* Interceptor configuration: Added `"lto1"` to the list of supported
binaries in the `Interceptor` setup to ensure diagnostics are correctly
intercepted for LTO (Link Time Optimization) builds, alongside `cc1` and
`cc1plus`.