fix(rt): backport fix from LLVM upstream
In ShadowSet, when p == end == ShadowMem::end, it triggered an
 assertion fail previously.

Now we do not allow p == end anymore in ShadowSet.

I have made a merged PR into LLVM upstream to fix such bug, but
forgotten to apply it into our XSan, as follows:
https://github.com/llvm/llvm-project/pull/144994/files

Now, we backport the relevant fix to XSan.